backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
pandas
python-dotenv
